Move PushDownContext outside of the AbstractCalciteIndexScan
